fix dataset dtype

+++ data.py
@@ -19,7 +19,7 @@ class Dataset(TorchDataset):
         # Normalize based on pre-computed mean and std of training set
         data = (data - [[[-10.4053]], [[-17.2822]], [[38.7486]]]) / [[[4.0457]], [[4.7750]], [[33.6621]]]
 
-        return data, label[None, :, :]
+        return data.astype(np.float32), label[None, :, :]
 
 
 class TrainDataset(Dataset):
